% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scGPSFuncs.r
\name{annotate_scGPS}
\alias{annotate_scGPS}
\title{annotate_scGPS}
\usage{
annotate_scGPS(DEgeneList, pvalueCutoff = 0.05, gene_symbol = TRUE,
  species = "human", universe)
}
\arguments{
\item{DEgeneList}{the list of genes being interrogated}

\item{pvalueCutoff}{self explanatory default 0.05}

\item{gene_symbol}{logical}

\item{species}{character default "human"}

\item{universe}{my addiditon, a vector of symbols representing the background}
}
\value{
a reactome pathway analysis
}
\description{
had to modify it coz the native scGPS function doesnt allow for a background gene list
}
\examples{
NULL

}
