% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFuncs.R
\name{annotateByCuts}
\alias{annotateByCuts}
\title{annotateByCuts}
\usage{
annotateByCuts(seuratObj, geneTab, expCut, groupBy, groupID, splitBy,
  splitID = NULL)
}
\arguments{
\item{seuratObj}{the Seurat object}

\item{geneTab}{the a data.frame of gene and cuts}

\item{groupBy}{the metadata to group by (like cell type)}

\item{groupID}{a pattern matching group under investigation, required}

\item{splitBy}{the metadata to split by (like experiment)}

\item{splitID}{a subset of splitBy if required, default NULL (all)}
}
\value{
Seurat Object with calls and outTab, a summary of the results
}
\description{
takes a gene and a cutoff and counts the number of cells in group expressing it
}
\examples{
tab<-data.frame(gene=c("PFN1","GZMA","GZMB"),cuts=c(2,0,0))
tmp<-annotateByCuts(seuratObj=integrated,geneTab=tab,groupBy="orig_final.ident",groupID="CD8",splitBy="orig.ident")

}
