% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFuncs.R
\name{buildMasterSeurat}
\alias{buildMasterSeurat}
\title{buildMasterSeurat}
\usage{
buildMasterSeurat(seuratList, ndims = 1:30, mt.cutoff = 10,
  regress.cell.cycle = TRUE, res.vec = c(seq(0, 1, 0.1), 1.5, 2))
}
\arguments{
\item{seuratList}{the list of Seurat objects, usually from makeSeuratList}

\item{ndims}{dimensions to use in integration,SCT and UMAP, default 1:30}

\item{mt.cutoff}{numeric percentage cutoff for mito expression, defaullt 10}

\item{regress.cell.cycle}{logical, whether to regress out cell cycle along with mito and nFeature}

\item{res.vec}{vector of resolutions to find clusters at, defaults to c(seq(0,1,0.1),1.5,2)}
}
\value{
an integrated master seurat object
}
\description{
turns a list of Seurat objects into a master object
}
\examples{
NULL

}
