% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFuncs.R
\name{countByCut}
\alias{countByCut}
\title{countByCut}
\usage{
countByCut(seuratObj, geneName, expCut, groupBy, splitBy)
}
\arguments{
\item{seuratObj}{the Seurat object}

\item{geneName}{the gene of interest}

\item{expCut}{the cutoff for high/low}

\item{groupBy}{the metadata to group by}

\item{splitBy}{the metadata to split by}
}
\value{
table
}
\description{
takes a gene and a cutoff and counts the number of cells in group expressing it
}
\examples{
countTab<-countByCut(seuratObj=integrated,geneName="PFN1",expCut=2,groupBy="orig_final.ident",splitBy="orig.ident")

}
