% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readInFuncs.R
\name{ExtractField}
\alias{ExtractField}
\title{helper function for Read10X}
\usage{
ExtractField(string, field = 1, delim = "_")
}
\arguments{
\item{string}{i dont know, ou just need it}
}
\value{
a vector of ids
}
\description{
turns barcodes.csv into "useful" ids
}
\examples{
NULL

}
