% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFuncs.R
\name{getEntrez}
\alias{getEntrez}
\title{getEntrez}
\usage{
getEntrez(seuratObj, geneList = "var.genes")
}
\arguments{
\item{seuratObj}{the Seurat object}

\item{geneList}{list of genes to calculate magic for default variable genes from SCT}
}
\value{
a dataframe of gene symbol and entrezID
}
\description{
use biomart to map gene symbols from Seurat object to entrez ids
}
\examples{
hasEntrez<-getEntrez(integrated)

}
