% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFuncs.R
\name{intgraph}
\alias{intgraph}
\title{intgraph}
\usage{
intgraph(cellphoneDB_data, scoreCut = 0.3, numberCut = 0,
  numberSplit = 35)
}
\arguments{
\item{cellphoneDB_data}{the countdat output from plotdat object}

\item{scoreCut}{mean score cutoff to use (seq(0,1,0.1)), default 0.3}

\item{numberCut}{the number of interactions required to be plotted, default 0}

\item{numberSplit}{the number of interactions to distinguish high from low, default 35}
}
\value{
invisible list of plot objects
}
\description{
takes a preprocCellphone$countdat object and plots a graph
}
\examples{
intgraph(x$countdat)

}
