% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFuncs.R
\name{makeMetaScore}
\alias{makeMetaScore}
\title{makeMetaScore}
\usage{
makeMetaScore(seuratObj, geneList, reduction = NA)
}
\arguments{
\item{seuratObj}{the Seurat object}

\item{geneList}{list of genes to calculate magic for default to a handfull of EMT genes}
}
\value{
a dataframe of metadata +/- coords and metascore
}
\description{
adds a gene metascore to meta.data
}
\examples{
integrated<-makeMetaScore(integrated,sample(rownames(integrated@assays$SCT@scale.data),10))
ggplot(integrated@meta.data,aes(x=UMAP_1,y=UMAP_2,color=metascore)) + geom_point(alpha=.5) + scale_colour_gradientn(colours = jet.colors(7))

}
