% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFuncs.R
\name{makeReactomeForMarkers}
\alias{makeReactomeForMarkers}
\title{makeReactomeForMarkers}
\usage{
makeReactomeForMarkers(findMarkersOut, hasEntrezObj)
}
\arguments{
\item{findMarkersOut}{a typical output from a FindMarkers seurat operation}

\item{hasEntrezObj}{an output from getEntrez on the seurat object}
}
\value{
list with markers and a list of compareClusterObject for plotting
}
\description{
wrapper for output of FindClusters to reactome but will crap out if no significance
}
\examples{
enrichment<-makeSpecificMarkers(integrated,groupBy_1="orig_final.ident",groupID_1="Cancer_2",groupBy_2="PFN1_GZMA_GZMB",groupID_2=c("hi_Low_Low","Low_Low_Low"),splitBy="orig.ident",splitID=NULL)
out<-makeReactomeForMarkers(enrichment$markerLists[[1]],enrichment$hasEntrez)

}
