% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFuncs.R
\name{makeReactomePipe}
\alias{makeReactomePipe}
\title{makeReactomePipe}
\usage{
makeReactomePipe(seuratObj, metaDataColumn)
}
\arguments{
\item{seuratObj}{an a seurat object}

\item{metaDataColumn}{a column name from seuratObj metadata}
}
\value{
list with markers and a list of compareClusterObject for plotting
}
\description{
wrapper for seurat to reactome
}
\examples{
enrichment<-makeReactomePipe(integrated,"SCT_snn_res.0.15")
for i in 1:length(enrichment$CP_result)) dotplot(enrichment$CP_result[[i]]) + ggtitle(names(enrichment$CP_result)[i])

}
