% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFuncs.R
\name{makeSpecificMarkers}
\alias{makeSpecificMarkers}
\title{makeSpecificMarkers}
\usage{
makeSpecificMarkers(integrated, groupBy_1, groupID_1, groupBy_2, groupID_2,
  splitBy, splitID = NULL)
}
\arguments{
\item{groupBy_1}{the metadata to group by (like cell type)}

\item{groupID_1}{a specific value of groupBy, required}

\item{groupBy_2}{the metadata to group by (like a result from annotateByCuts)}

\item{groupID_2}{two values of groupBy_2, required}

\item{splitBy}{the metadata to split by (like experiment)}

\item{splitID}{a subset of splitBy if required, default NULL (all)}

\item{mkEntObj}{an output from  mkEnt or an element of splotDirection}

\item{seuratObj}{the Seurat object}
}
\value{
list with markers and a list of compareClusterObject for plotting
}
\description{
wrapper for seurat to reactome but identifying two metadata columns and specifying a single 1vs1 comparison (splits by experiment)
}
\examples{
enrichment<-makeSpecificMarkers(integrated,groupBy_1="orig_final.ident",groupID_1="Cancer_2",groupBy_2="PFN1_GZMA_GZMB",groupID_2=c("hi_Low_Low","Low_Low_Low"),splitBy="orig.ident",splitID=NULL)
for i in 1:length(enrichment$CP_result)) dotplot(enrichment$CP_result[[i]]) + ggtitle(names(enrichment$CP_result)[i])

}
