% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFuncs.R
\name{mkEnt}
\alias{mkEnt}
\title{mkEnt}
\usage{
mkEnt(listOids, getEntrezObj)
}
\arguments{
\item{listOids}{a list of gene symbol vectors}

\item{getEntrezObj}{lan output from getEnrez functin}
}
\value{
a list of entrez id vectors for enrichment
}
\description{
retrieves entrez ids from a getEntrez for a list of vectors of gene symbols
}
\examples{
entList<-mkEnt(split(markerList$gene,markerList$cluster),hasEntrez)

}
