% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scGPSFuncs.r
\name{mklasso}
\alias{mklasso}
\title{mklasso}
\usage{
mklasso(scPredObj)
}
\arguments{
\item{scPredObj}{an output of scPred}
}
\value{
LASSO object that can be tabulated or plotted
}
\description{
turns a scpred object into a LASSO object
}
\examples{
NULL

}
