% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFuncs.R
\name{mkUpset}
\alias{mkUpset}
\title{mkUpset}
\usage{
mkUpset(categoryMatrix, plotAll = T, ...)
}
\arguments{
\item{plotAll}{logical, plot all possible comparisons (even empty ones), default TRUE}

\item{...}{additional parameters to pass to upset if plotall is set to FALSE}

\item{seuratObj}{the takes a dataframe of categories}
}
\value{
an upset plot in working directory called upsetPlot.pdf
}
\description{
takes a dataframe of categories and makes an upset plot
}
\examples{
NULL
mkUpset(integrated@meta.data[,c('PFN1',"GZMA","GZMB")],plotAll=F)
}
