% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scGPSFuncs.r
\name{plotLasso}
\alias{plotLasso}
\title{plotLasso}
\usage{
plotLasso(mklassoObj)
}
\arguments{
\item{mklassoObj}{an output of mklasso}
}
\value{
LASSO object that can be tabulated or plotted
}
\description{
turns a mklasso object into a sankey diagram
}
\examples{
NULL

}
