% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFuncs.R
\name{plotSankey}
\alias{plotSankey}
\title{plotSankey}
\usage{
plotSankey(seuratObj, idvar = c("varRes.0.3", "emt_res.0.3"))
}
\arguments{
\item{seuratObj}{the Seurat object}

\item{idvars}{character vector of medatadata column names}
}
\value{
a gawd awful javascript doohicky that i dont understand
}
\description{
compares populations between metadata columns
}
\examples{
x<-plotSankey(integrated,c("orig.ident","SCT_snn_res.0.8"))

}
