% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monocleFuncs.R
\name{plotTraj}
\alias{plotTraj}
\title{plotTraj}
\usage{
plotTraj(monocleObj, geneVec)
}
\arguments{
\item{monocleObj}{the monocle object}

\item{geneVec}{the character vector of genes to assemble cells on}
}
\value{
a monocle object with Pseudo vectors and state in me
}
\description{
turns a monocle object into a monocle object ordered by a gene list
}
\examples{
NULL

}
