% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFuncs.R
\name{preprocCellphone}
\alias{preprocCellphone}
\title{preprocCellPhone}
\usage{
preprocCellphone(prefix, pval = 0.01, varval = 0)
}
\arguments{
\item{pval}{p value cutoff for reporting an interaction, default 0.01}

\item{varval}{variance value cutoff for retaining interactions that are different between conditions, default 0}

\item{dir}{the directory containing cellPhoneDB output, defaults to .}
}
\value{
a list containing raw data trimmed for 0 and one trimmed for p and var (plotdat)
}
\description{
takes CellPhoneDB output and turns it into something useable
}
\examples{
cd("PATH/TO/CELLPHONEDB/OUT/small")
x<-preprocCellphone(varval=0,pval=.05)

}
