% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFuncs.R
\name{reactomeClusts}
\alias{reactomeClusts}
\title{reactomeClusts}
\usage{
reactomeClusts(mkEntObj)
}
\arguments{
\item{mkEntObj}{an output from  mkEnt or an element of splotDirection}
}
\value{
compareClusterObject for plotting
}
\description{
wrapper for compareCluster reactome for a list of entrez ID vectors
}
\examples{
oi<-reactomeClusts(mkEntObj)
maybe dotplot(oi)
oi<-lapply(splitDirectionObj,reactomeClusts)
dotplot(oi[[1]])

}
