% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readInFuncs.R
\name{Read10X}
\alias{Read10X}
\title{Modified read 10x function}
\usage{
Read10X(data.dir = NULL)
}
\arguments{
\item{data.dir}{a directory path containing cellranger output}
}
\value{
a matrix of counts
}
\description{
Made when the on board function in seurat was broken
}
\examples{
raw.data <- Read10X("PATH/TO/OUT")

}
