% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readInFuncs.R
\name{read10xCounts}
\alias{read10xCounts}
\title{read10xCounts}
\usage{
read10xCounts(samples, col.names = FALSE, ...)
}
\arguments{
\item{samples}{a directory containing cellranger output}
}
\value{
a SCE object
}
\description{
modified version of SCE function to accept features.tsc
}
\examples{
NULL

}
