% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readInFuncs.R
\name{read10xMatrix}
\alias{read10xMatrix}
\title{read10xMatrix}
\usage{
read10xMatrix(file, hdf5.out = FALSE, chunk.size)
}
\arguments{
\item{file}{from 10xcounts}
}
\value{
a SCE object
}
\description{
handed to read10xcounts
}
\examples{
NULL

}
