% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCEfuncs.R
\name{SCEpipe}
\alias{SCEpipe}
\title{SCEpipe}
\usage{
SCEpipe(SCElist, countsProp = 0.8, mitoProp = 0.1)
}
\arguments{
\item{SCElist}{a list of SingleCellExperiment objects from a species mixing experiment}

\item{countsProp}{cut off of minimum proportion of counts in human, default 0.8}

\item{mitoProp}{cut off of maximum  proportion of mito counts in human, default 0.1}
}
\value{
a list of trimmed and cleaned
}
\description{
Annotate mouse and human, plus mito, trim, calculate QC for a list of SCE object
}
\examples{
NULL

}
