% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFuncs.R
\name{seurat2monocle}
\alias{seurat2monocle}
\title{seurat2monocle}
\usage{
seurat2monocle(seuratObj)
}
\arguments{
\item{seuratObj}{the Seurat object}
}
\value{
a monocle object
}
\description{
turns a Seurat object into a monocle object
}
\examples{
NULL

}
