% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFuncs.R
\name{seurat2scGPS}
\alias{seurat2scGPS}
\title{seurat2scGPS}
\usage{
seurat2scGPS(seuratObj)
}
\arguments{
\item{seuratObj}{the Seurat object}
}
\value{
a scGPS object
}
\description{
turns a Seurat object into a scGPS object
}
\examples{
NULL

}
