% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFuncs.R
\name{splitDirection}
\alias{splitDirection}
\title{splitDirection}
\usage{
splitDirection(marker)
}
\arguments{
\item{marker}{a Find[All]Markers object (may need marker$gene to be added)}
}
\value{
a matrix of MAGIC expression
}
\description{
turns a  Seurat FindMarkers object into a bunch of entrez IDs
}
\examples{
NULL

}
