% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFuncs.R
\name{splitMasterSeurat}
\alias{splitMasterSeurat}
\title{splitMasterSeurat}
\usage{
splitMasterSeurat(seuratObj, metaData, keep = NULL, ndims = 1:30,
  resVec = c(seq(0, 1, 0.1), 1.5, 2))
}
\arguments{
\item{seuratObj}{the list of Seurat objects, usually from makeSeuratList}

\item{metaData}{the metadata column you want to split on}

\item{keep}{a character vector of a subset of metadata that you want to keep, default NULL}

\item{ndims}{dimensions to use in integration,SCT and UMAP, default 1:30}

\item{res.vec}{vector of resolutions to find clusters at, defaults to c(seq(0,1,0.1),1.5,2)}
}
\value{
an list of seurat objects
}
\description{
turns a Seurat objects into a list of seurat objects
}
\examples{
NULL

}
