% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCEfuncs.R
\name{trimNscent}
\alias{trimNscent}
\title{trimNscent}
\usage{
trimNscent(sce_obj)
}
\arguments{
\item{sce_obj}{a SingleCellExperiment object}
}
\value{
invisible list of plot objects
}
\description{
takes a singleCellExperiment object and runs landSCENT
}
\examples{
NULL

}
