% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFuncs.R
\name{makeReactomeForMarkers}
\alias{makeReactomeForMarkers}
\title{makeReactomeForMarkers}
\usage{
makeReactomeForMarkers(findMarkersOut, hasEntrezObj)
}
\arguments{
\item{findMarkersOut}{a typical output from a FindMarkers seurat operation}

\item{hasEntrezObj}{an output from getEntrez on the seurat object}
}
\value{
list with markers and a list of compareClusterObject for plotting
}
\description{
wrapper for output of FindClusters to reactome but will crap out if no significance
}
\examples{
 enrichment<-makeSpecificMarkers(integrated,groupBy_1="orig_final.ident",groupID_1="CD8 T-cell 1",groupBy_2="orig.ident",groupID_2=c("ACITE","BCITE"),splitBy=NULL,splitID=NULL)
out<-makeReactomeForMarkers(enrichment$markerLists,enrichment$hasEntrez)
dotoplot(out)

}
