% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFuncs.R
\name{makeSpecificMarkers}
\alias{makeSpecificMarkers}
\title{makeSpecificMarkers}
\usage{
makeSpecificMarkers(seuratObj, groupBy_1, groupID_1, groupBy_2, groupID_2,
  splitBy = NULL, splitID = NULL)
}
\arguments{
\item{seuratObj}{the Seurat object}

\item{groupBy_1}{the metadata to group by (like cell type)}

\item{groupID_1}{a specific value of groupBy, required}

\item{groupBy_2}{the metadata to group by (like a result from annotateByCuts)}

\item{groupID_2}{two values of groupBy_2, required}

\item{splitBy}{optional; the metadata to split by (e.g. experiment), default NULL}

\item{splitID}{a subset of splitBy if splitBy is declared}

\item{mkEntObj}{an output from  mkEnt or an element of splotDirection}
}
\value{
list with markers and a list of compareClusterObject for plotting
}
\description{
wrapper for seurat to reactome but identifying two metadata columns and specifying a single 1vs1 comparison (splits by experiment)
}
\examples{
enrichment<-makeSpecificMarkers(integrated,groupBy_1="orig_final.ident",groupID_1="CD8 T-cell 1",groupBy_2="orig.ident",groupID_2=c("ACITE","BCITE"),splitBy=NULL,splitID=NULL)

}
